reshuffle some code to source files, reduce dependencies
