+++ Core/CMakeLists.txt
@@ -114,6 +114,7 @@ add_subdirectory(src/Detector)
 add_subdirectory(src/Geometry)
 add_subdirectory(src/MagneticField)
 add_subdirectory(src/Material)
+add_subdirectory(src/Navigation)
 add_subdirectory(src/Propagator)
 add_subdirectory(src/Surfaces)
 add_subdirectory(src/TrackFinding)
