+++ Core/include/Acts/Detector/DetectorVolume.hpp
@@ -432,122 +432,5 @@ class DetectorVolumeFactory {
   }
 };
 
-/// Helper extractors: all portals
-struct AllPortalsExtractor {
-  /// Extract the portals from the volume
-  ///
-  /// @param gctx the geometry contextfor this extraction call
-  /// @param nState is the current navigation state
-  ///
-  /// @return a vector of raw Portal pointers
-  inline static const std::vector<const Portal*> extract(
-      [[maybe_unused]] const GeometryContext& gctx,
-      const NavigationState& nState) {
-    if (nState.currentVolume == nullptr) {
-      throw std::runtime_error(
-          "AllPortalsExtractor: no detector volume given.");
-    }
-    return nState.currentVolume->portals();
-  }
-};
-
-/// Helper extractors: all surfaces
-struct AllSurfacesExtractor {
-  /// Extract the surfaces from the volume
-  ///
-  /// @param gctx the geometry contextfor this extraction call
-  /// @param nState is the current navigation state
-  /// @param indices is an ignored index vector
-  ///
-  /// @return a vector of raw Surface pointers
-  inline static const std::vector<const Surface*> extract(
-      [[maybe_unused]] const GeometryContext& gctx,
-      const NavigationState& nState,
-      [[maybe_unused]] const std::vector<size_t>& indices = {}) {
-    if (nState.currentVolume == nullptr) {
-      throw std::runtime_error(
-          "AllSurfacesExtractor: no detector volume given.");
-    }
-    return nState.currentVolume->surfaces();
-  }
-};
-
-/// Helper extractors: indexed surfaces
-struct IndexedSurfacesExtractor {
-  /// Extract the surfaces from the volume
-  ///
-  /// @param gctx the geometry contextfor this extraction call
-  /// @param nState is the current navigation state
-  /// @param indices are access indices into the surfaces store
-  ///
-  /// @note no out of boudns checking is done
-  ///
-  /// @return a vector of raw Surface pointers
-  inline static const std::vector<const Surface*> extract(
-      [[maybe_unused]] const GeometryContext& gctx,
-      const NavigationState& nState, const std::vector<size_t>& indices) {
-    if (nState.currentVolume == nullptr) {
-      throw std::runtime_error(
-          "IndexedSurfacesExtractor: no detector volume given.");
-    }
-    // Get the surface container
-    const auto& surfaces = nState.currentVolume->surfaces();
-    // The extracted surfaces
-    std::vector<const Surface*> eSurfaces;
-    eSurfaces.reserve(indices.size());
-    std::for_each(indices.begin(), indices.end(),
-                  [&](const auto& i) { eSurfaces.push_back(surfaces[i]); });
-    return eSurfaces;
-  }
-};
-
-/// Helper extractors: all sub volumes of a volume
-struct AllSubVolumesExtractor {
-  /// Extract the sub volumes from the volume
-  ///
-  /// @param gctx the geometry contextfor this extraction call
-  /// @param nState is the current navigation state
-  /// @param indices are access indices into the volume store (ignored)
-  ///
-  /// @return a vector of raw DetectorVolume pointers
-  inline static const std::vector<const DetectorVolume*> extract(
-      [[maybe_unused]] const GeometryContext& gctx,
-      const NavigationState& nState,
-      [[maybe_unused]] const std::vector<size_t>& indices = {}) {
-    if (nState.currentVolume == nullptr) {
-      throw std::runtime_error(
-          "AllSubVolumesExtractor: no detector volume given.");
-    }
-    return nState.currentVolume->volumes();
-  }
-};
-
-/// Helper extractors: indexed sub volume of a volume
-struct IndexedSubVolumesExtractor {
-  /// Extract the sub volumes from the volume
-  ///
-  /// @param gctx the geometry contextfor this extraction call
-  /// @param nState is the current navigation state
-  /// @param indices are access indices into the volume store
-  ///
-  /// @return a vector of raw DetectorVolume pointers
-  inline static const std::vector<const DetectorVolume*> extract(
-      [[maybe_unused]] const GeometryContext& gctx,
-      const NavigationState& nState, const std::vector<size_t>& indices) {
-    if (nState.currentVolume == nullptr) {
-      throw std::runtime_error(
-          "AllSubVolumesExtractor: no detector volume given.");
-    }
-    // Get the sub volumes container
-    const auto& volumes = nState.currentVolume->volumes();
-    // The extracted volumes
-    std::vector<const DetectorVolume*> eVolumes;
-    eVolumes.reserve(indices.size());
-    std::for_each(indices.begin(), indices.end(),
-                  [&](const auto& i) { eVolumes.push_back(volumes[i]); });
-    return eVolumes;
-  }
-};
-
 }  // namespace Experimental
 }  // namespace Acts
