+++ Core/include/Acts/Detector/detail/PortalHelper.hpp
@@ -10,8 +10,8 @@
 
 #include "Acts/Definitions/Algebra.hpp"
 #include "Acts/Definitions/Common.hpp"
-#include "Acts/Detector/DetectorComponents.hpp"
 #include "Acts/Definitions/Direction.hpp"
+#include "Acts/Detector/DetectorComponents.hpp"
 #include "Acts/Detector/Portal.hpp"
 #include "Acts/Geometry/GeometryContext.hpp"
 #include "Acts/Utilities/BinningType.hpp"
