+++ Core/src/Navigation/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(
+  ActsCore
+  PRIVATE
+  SurfaceCandidatesUpdators.cpp
+)
